requirements.txt
appdirs==1.4.4
attrs==21.2.0
autopep8==1.6.0
cattrs==1.9.0
certifi==2021.10.8
charset-normalizer==2.0.9
etherscan-python==2.1.0
idna==3.3
numpy==1.21.5
pandas==1.3.5
pycodestyle==2.8.0
pycoingecko==2.2.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-cache==0.8.1
six==1.16.0
toml==0.10.2
url-normalize==1.4.3
urllib3==1.26.7
